One Front Door/SourceCode/DPSG.Web.Merchandiser/src/WebApp/src/app/planning/StoreSetup/storelist-popup/storelist-popup.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { StorelistPopupComponent } from './storelist-popup.component';

describe('Component: StorelistPopup', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [StorelistPopupComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([StorelistPopupComponent],
      (component: StorelistPopupComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(StorelistPopupComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(StorelistPopupComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-storelist-popup></app-storelist-popup>
  `,
  directives: [StorelistPopupComponent]
})
class StorelistPopupComponentTestController {
}

